15_opentelemetry_tracing_ml_inference/requirements.txt
fastapi
uvicorn[standard]
httpx
sqlalchemy
asyncpg
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-logging
